Added Andrew's Changes and Some more Silk Screen ID
Hi Taylor,

> We've been getting great feedback from you on our schematic/layout and
I
> was wondering if you could take one last look at our layout to see if
we
> missed anything major before we submit it.

I didn't have much time to look, but here you go:

Schematic:

- IMPORTANT: Chip "74ACT125" has an uninvoked gate. It's a quad buffer
part, and one of the buffers isn't hooked up. Use the INVOKE command in
eagle and click on that part to get the fourth buffer. Tie the input and
the enable line to low/high as you see fit -- but don't leave it
floating, like it is now.

- IMPORTANT: Chip "74ACT125" needs aq bypass cap.

- NOT IMPORTANT: Usually schematics have part identifiers that have
non-semantic meanings: like R1, R2, R3 not "PULL_UP_1". This is because
things move around a lot, and it's easier to type, and not get mixed up.
That said, I don't really care, feel free to leave it that way. But in
general, with schematics you share with others, I'd stick with regular
part identifiers.

Layout:

- MAYBE IMPORTANT: your ISP connector is a little close to the chip
below it and conector above. I think you'll be OK, but I'm not sure -
you might want to move that other connector a bit up away from the ISP
connector. But if you've printed it out and it looks OK, ignore this.

- NOT IMPORTANT: your board mounting holes are 0.115 inches, which is a
REALLY tight fit for a #4 screw. You might want to open it up to 0.125
if you have room.

- Otherwise: Looks great!